+++ CHANGELOG.md
@@ -9,12 +9,15 @@ and the project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- Sanitize molecules from GraphGA ([#5](https://github.com/AustinT/mol_ga/pull/5)) ([@austint])
+
 ### Added
 
 ### Fixed
 
 - Fix import error for python<3.8 ([#3](https://github.com/AustinT/mol_ga/pull/3)) ([@austint])
 - Fix unintended use of system random in sampling ([#4](https://github.com/AustinT/mol_ga/pull/4)) ([@austint])
+- Fix occasional crash from rdkit errors during crossover ([#5](https://github.com/AustinT/mol_ga/pull/5)) ([@austint])
 
 ## [0.1.0] - 2023-09-05